% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{build_URL}
\alias{build_URL}
\title{Build URL}
\usage{
build_URL(prodnum, filters = "")
}
\arguments{
\item{prodnum}{The product number. Passed from product function.}
}
\description{
This function builds the .json product URL.
}
\examples{
build_URL()
}
