% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madisonR.R
\name{carbon_footprint}
\alias{carbon_footprint}
\title{Reports the carbon footprint from the known ingredients}
\usage{
carbon_footprint(item)
}
\arguments{
\item{item}{The output of a specific product() call}
}
\value{
A 'double' with amount of carbon footprint, in g
}
\description{
Reports the carbon footprint from the known ingredients
}
\examples{
carbon_footprint(product("cracker"))
carbon_footprint(cracker) # where cracker is saved variable from product() call
}
