% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{check_internet}
\alias{check_internet}
\title{Check internet.}
\usage{
check_internet()
}
\description{
This function verifies that you have an internet connection. For internal use only.
}
\examples{
check_internet()
}
