% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\docType{data}
\name{countries}
\alias{countries}
\title{Valid countries and their codes.}
\format{
A data frame with 2 variables:
\itemize{
\item names: full country name
\item codes: 2-letter country code
}
}
\usage{
countries
}
\description{
A 2-variable dataframe. The variables are as follows:
}
\keyword{datasets}
