% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickys.R
\name{food_group}
\alias{food_group}
\title{Report the food group product is in}
\usage{
food_group(item)
}
\arguments{
\item{item}{The output of a specific product() call}
}
\value{
A string with the group
}
\description{
Report the food group product is in
}
\examples{
food_group(product("cookie"))
food_group(cookie) # where cookie is saved variable from product() call
}
