% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madisonR.R
\name{plot_carbon}
\alias{plot_carbon}
\title{Bar Plot comparing the carbon footprint (in g) of multiple products}
\usage{
plot_carbon(prod_list)
}
\arguments{
\item{prod_list}{A list of pre-defined products}
}
\value{
The plot of product names vs carbon footprint
}
\description{
Bar Plot comparing the carbon footprint (in g) of multiple products
}
\examples{
plot_carbon(list(cracker, apple, cheese, countries)) # saved variables
plot_carbon(list(product("cracker"), product("apple")))
}
