% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickys.R
\name{plot_sugar}
\alias{plot_sugar}
\title{Bar Plot of the amount of sugar in a list of products}
\usage{
plot_sugar(prod_list)
}
\arguments{
\item{prod_list}{A list of pre-defined products}
}
\value{
The plot of product names vs amount of sugar
}
\description{
Bar Plot of the amount of sugar in a list of products
}
\examples{
plot_sugar(list(cookie, apple)) # where cookie and apple are saved variables from product() calls
plot_sugar(list(product("cookie"), product("apple")))
}
