% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickys.R
\name{prod_name}
\alias{prod_name}
\title{Report the product name}
\usage{
prod_name(item)
}
\arguments{
\item{item}{The output of a specific product() call}
}
\value{
A string with the product name
}
\description{
This function returns the product name of a json from product() function
}
\examples{
prod_name(product("cookie"))
prod_name(cookie) # where cookie is saved variable from product() call
}
