% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R, R/jonahs.R
\name{product}
\alias{product}
\title{Get a product.}
\usage{
product(term, chars = 30, num = NA, country = "world")

product(term, chars = 30, num = NA, country = "world")
}
\arguments{
\item{term}{The name of the product to search for.}

\item{chars}{The number of characters from the product name to print.}

\item{num}{The product number (1:24) to retrieve.}

\item{country}{The country code. Searches only for products in the given country.}
}
\value{
A nested list of product information, parsed from the JSON that is returned by the API.
}
\description{
This function retrieves a .json file for a product.

This function retrieves a .json file for a product.
}
\examples{
product('eggs')
product()
}
