% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R, R/jonahs.R
\name{product_by_prodnum}
\alias{product_by_prodnum}
\title{Product by Product Number}
\usage{
product_by_prodnum(prodnum)

product_by_prodnum(prodnum)
}
\arguments{
\item{prodnum}{The product number. Passed from product function.}
}
\description{
If you already know the product number of the product you want, you can use this function to retrieve the .json.

If you already know the product number of the product you want, you can use this function to retrieve the .json.
}
\examples{
product_by_prodnum()
product_by_prodnum()
}
