% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R, R/jonahs.R
\name{retrieve_countries}
\alias{retrieve_countries}
\title{Print countries}
\usage{
retrieve_countries()

retrieve_countries()
}
\description{
This function prints the available countries as well as their country codes.

This function prints the available countries as well as their country codes.
}
\examples{
retrieve_countries()
retrieve_countries()
}
