% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R, R/jonahs.R
\name{search_by_name}
\alias{search_by_name}
\title{Search Product by Name}
\usage{
search_by_name(term, country = "world")

search_by_name(term, country = "world")
}
\arguments{
\item{term}{The name of the product to search for.}

\item{country}{The country code. Searches only for products in the given country.}
}
\description{
This function allows you to express your love of cats.

This function allows you to express your love of cats.
}
\examples{
search_by_name()
search_by_name()
}
