% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickys.R
\name{sugar_per_100g}
\alias{sugar_per_100g}
\title{Report the amount of sugar in product per 100g}
\usage{
sugar_per_100g(item)
}
\arguments{
\item{item}{The output of a specific product() call}
}
\value{
A 'double' with amount of sugar
}
\description{
Report the amount of sugar in product per 100g
}
\examples{
sugar_per_100g(product("cookie"))
sugar_per_100g(cookie) # where cookie is saved variable from product() call
}
