% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madisonR.R
\name{to_csv}
\alias{to_csv}
\title{Convert product json into CSV / Excel}
\usage{
to_csv(item)
}
\arguments{
\item{item}{The output of a specific product() call}
}
\value{
A csv with product info
}
\description{
This function converts the product info from json to csv from product() function
}
\examples{
to_csv(product("cracker"))
to_csv(cracker) # where cracker is saved variable from product() call
}
