% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{update_countries}
\alias{update_countries}
\title{Update countries.}
\usage{
update_countries()
}
\description{
This function updates the list of available countries using the API.
}
\examples{
update_countries()
}
