% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madisonR.R
\name{view_allergens}
\alias{view_allergens}
\title{Viewing list of allergens of a product}
\usage{
view_allergens(item)
}
\arguments{
\item{item}{The output of a specific product() call}
}
\value{
A list with the product allergens
}
\description{
This function returns a list of allergens from product() function
}
\examples{
view_allergens(product("cracker"))
view_allergens(cracker) # where cracker is saved variable from product() call
}
