% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{build_URL}
\alias{build_URL}
\title{Build URL}
\usage{
build_URL(prodnum, country = "world", filters = "")
}
\arguments{
\item{prodnum}{The product number. Passed from product function.}

\item{country}{The country code. Changes the language of the .json file.}
}
\description{
This function builds the .json product URL.
}
\examples{
build_URL()
}
