% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{product}
\alias{product}
\title{Get a product.}
\usage{
product(term, chars = 30)
}
\arguments{
\item{term}{The name of the product to search for.}
}
\description{
This function retrieves a .json file for a product.
}
\examples{
product()
}
