% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R, R/jonahs.R
\name{update_countries}
\alias{update_countries}
\title{Update countries.}
\usage{
update_countries()

update_countries()
}
\description{
This function updates the list of available countries using the API. Warning: resulting 'countries' variable is written to the global env.

This function updates the list of available countries using the API. Warning: resulting 'countries' variable is written to the global env.
}
\examples{
update_countries()
update_countries()
}
